split app.py
