+++ config.py
@@ -1,0 +1,19 @@
+# SQLALCHEMY
+SQLALCHEMY_DATABASE_URI = 'mysql+pymysql://root:rootpass@localhost/taipei_day_trip'
+
+# session
+SECRET_KEY = 'eovnlfnlonlkjmflo'
+
+# database
+DATABASE_HOST = 'localhost'
+DATABASE_NAME = 'taipei_day_trip'
+DATABASE_USER = 'root'
+DATABASE_PASSWORD = 'rootpass'
+
+# connection pool
+CONNECTION_POOL_SIZE = 5
+
+# 其他
+JSON_AS_ASCII = False
+TEMPLATES_AUTO_RELOAD = True
+JSON_SORT_KEYS = False
